[ember-data] Move type registries into their own ghost modules
