+++ types/ember-data/adapter.d.ts
@@ -1,3 +1,3 @@
 import DS from 'ember-data';
 export default DS.Adapter;
-export { AdapterRegistry } from 'ember-data';
+export { default as AdapterRegistry } from 'ember-data/types/registries/adapter';
