+++ types/ember-data/model.d.ts
@@ -1,3 +1,3 @@
 import DS from 'ember-data';
 export default DS.Model;
-export { ModelRegistry } from 'ember-data';
+export { default as ModelRegistry } from 'ember-data/types/registries/model';
