+++ types/ember-data/serializer.d.ts
@@ -1,3 +1,3 @@
 import DS from 'ember-data';
 export default DS.Serializer;
-export { SerializerRegistry } from 'ember-data';
+export { default as SerializerRegistry } from 'ember-data/types/registries/serializer';
