+++ types/ember-data/test/record-reference.ts
@@ -7,8 +7,8 @@ class User extends DS.Model {
     username = DS.attr('string');
 }
 
-declare module 'ember-data' {
-    interface ModelRegistry {
+declare module 'ember-data/types/registries/model' {
+    export default interface ModelRegistry {
         user: User;
     }
 }
