+++ types/ember-data/test/store.ts
@@ -10,8 +10,8 @@ class Post extends DS.Model {
     comments = DS.hasMany('comment');
 }
 
-declare module 'ember-data' {
-    interface ModelRegistry {
+declare module 'ember-data/types/registries/model' {
+    export default interface ModelRegistry {
         'post': Post;
         'post-comment': PostComment;
     }
@@ -39,8 +39,8 @@ class User extends DS.Model {
 
 class Author extends User {}
 
-declare module 'ember-data' {
-    interface ModelRegistry {
+declare module 'ember-data/types/registries/model' {
+    export default interface ModelRegistry {
         'user': User;
         'author': Author;
     }
@@ -73,8 +73,8 @@ class Message extends DS.Model {
     hasBeenSeen = DS.attr('boolean');
 }
 
-declare module 'ember-data' {
-    interface ModelRegistry {
+declare module 'ember-data/types/registries/model' {
+    export default interface ModelRegistry {
         message: Message;
     }
 }
@@ -191,12 +191,14 @@ class UserSerializer extends DS.Serializer {
     thisSerializerOnlyMethod(): void {}
 }
 
-declare module 'ember-data' {
-    interface AdapterRegistry {
+declare module 'ember-data/types/registries/adapter' {
+    export default interface AdapterRegistry {
         user: UserAdapter;
     }
+}
 
-    interface SerializerRegistry {
+declare module 'ember-data/types/registries/serializer' {
+    export default interface SerializerRegistry {
         user: UserSerializer;
     }
 }
