+++ types/ember-data/tsconfig.json
@@ -53,6 +53,10 @@
         "serializer.d.ts",
         "store.d.ts",
         "transform.d.ts",
+        "types/registries/adapter.d.ts",
+        "types/registries/model.d.ts",
+        "types/registries/serializer.d.ts",
+        "types/registries/transform.d.ts",
         "test/lib/assert.ts",
         "test/model.ts",
         "test/module-api.ts",
