+++ types/ember-data/types/registries/adapter.d.ts
@@ -1,0 +1,1 @@
+export default interface AdapterRegistry {}
