+++ types/ember-data/types/registries/model.d.ts
@@ -1,0 +1,3 @@
+interface ModelRegistry {}
+
+export default ModelRegistry;
