+++ types/ember-data/types/registries/serializer.d.ts
@@ -1,0 +1,1 @@
+export default interface SerializerRegistry {}
