+++ types/ember-data/types/registries/transform.d.ts
@@ -1,0 +1,6 @@
+export default interface TransformRegistry {
+    string: string;
+    boolean: boolean;
+    number: number;
+    date: Date;
+}
